+++ Sources/KankodaKit/About/AboutScreenContent.swift
@@ -80,6 +80,7 @@ private extension AboutScreenContent {
             icon: .lightbulb,
             description: Text("This app is developed by Kankoda in Stockholm, Sweden.\n\nDon't hesitate to reach out if you find bugs, or have ideas on how to improve the app.")
         )
+        .scrollContentBackground(.hidden)
         .navigationTitle("Test")
         #if os(iOS)
         .navigationBarTitleDisplayMode(.inline)