+++ Sources/KankodaKit/Vision/TextRecognitionContext.swift
@@ -6,7 +6,7 @@
 //  Copyright © 2023-2025 Kankoda. All rights reserved.
 //
 
-#if os(macOS) || os(iOS) || os(tvOS)
+#if os(macOS) || os(iOS) || os(tvOS) || os(visionOS)
 import SwiftUI
 import Vision
 
@@ -14,45 +14,139 @@ import Vision
 public class TextRecognitionContext: ObservableObject, @unchecked Sendable {
     
     public init() {}
-    
+
     @Published
-    public var result: [String] = []
-    
-    public func reset() { result = [] }
+    public var recognizedTexts: [String] = []
+
+    @Published
+    public var lastError: Error?
+
+    public func reset() {
+        lastError = nil
+        recognizedTexts = []
+    }
+}
+
+extension TextRecognitionContext {
+
+    func updateRecognizedTexts(
+        _ texts: [String]
+    ) async {
+        await MainActor.run {
+            lastError = nil
+            recognizedTexts = texts
+        }
+    }
+
+    func updateError(
+        _ err: Error
+    ) async {
+        await MainActor.run {
+            lastError = err
+            recognizedTexts = []
+        }
+    }
+}
+
+public enum TextRecognitionMethod: String, Equatable, Sendable {
+
+    case text, document
 }
 
 public extension TextRecognitionContext {
     
-    /// Perform text recognition in a collection of images.
-    func performTextRecognition(in images: [ImageRepresentable]) {
+    /// Perform text recognition in a set of images, using a
+    /// certain recognition method.
+    @available(iOS 26.0, macOS 26.0, tvOS 26.0, visionOS 26.0, *)
+    func performTextRecognition(
+        in images: [ImageRepresentable],
+        with method: TextRecognitionMethod = .document
+    ) {
         reset()
-        let images = images.compactMap { $0.cgImage }
-        for image in images {
-            let requestHandler = VNImageRequestHandler(cgImage: image, options: [:])
-            let request = self.textRecognitionRequest()
+        Task {
             do {
-                try requestHandler.perform([request])
+                let result = try await performInternal(in: images, with: method)
+                await updateRecognizedTexts(result)
             } catch {
-                print(error.localizedDescription)
+                await updateError(error)
             }
         }
     }
 }
 
-private extension TextRecognitionContext {
+extension TextRecognitionContext {
     
-    func textRecognitionRequest() -> VNRecognizeTextRequest {
-        let request = VNRecognizeTextRequest { request, error in
-            if let error = error { return print(error) }
-            guard let observations = request.results as? [VNRecognizedTextObservation] else { return }
-            let texts = observations.compactMap { $0.topCandidates(1).first?.string }
-            DispatchQueue.main.async {
-                self.result.append(contentsOf: texts)
-            }
+    /// Perform text recognition in a set of images, using a
+    /// certain recognition method.
+    @available(iOS 26.0, macOS 26.0, tvOS 26.0, visionOS 26.0, *)
+    func performInternal(
+        in images: [ImageRepresentable],
+        with method: TextRecognitionMethod = .document
+    ) async throws -> [String] {
+        switch method {
+        case .text:
+            let request = RecognizeTextRequest.kankodaRequest
+            return try await request.performTextRecognition(in: images)
+        case .document:
+            let request = RecognizeDocumentsRequest.kankodaRequest
+            return try await request.performTextRecognition(in: images)
         }
+    }
+}
+
+@available(iOS 26.0, macOS 26.0, tvOS 26.0, visionOS 26.0, *)
+extension RecognizeTextRequest {
+
+    static var kankodaRequest: Self {
+        var request = RecognizeTextRequest()
         request.recognitionLevel = .accurate
         request.usesLanguageCorrection = true
         return request
+    }
+
+    func performTextRecognition(
+        in images: [ImageRepresentable]
+    ) async throws -> [String] {
+        let images = images.compactMap { $0.cgImage }
+        var results: [String] = []
+        for image in images {
+            let observations = try await perform(on: image)
+            let texts = observations.compactMap {
+                $0.topCandidates(1).first?.string
+            }
+            results.append(contentsOf: texts)
+        }
+        return results
+    }
+}
+
+@available(iOS 26.0, macOS 26.0, tvOS 26.0, visionOS 26.0, *)
+extension RecognizeDocumentsRequest {
+
+    static var kankodaRequest: Self {
+        var request = RecognizeDocumentsRequest()
+        request.textRecognitionOptions.useLanguageCorrection = true
+        return request
+    }
+
+    enum KankodaError: LocalizedError {
+        case noDocumentFound
+    }
+
+    func performTextRecognition(
+        in images: [ImageRepresentable]
+    ) async throws -> [String] {
+        let images = images.compactMap { $0.cgImage }
+        var results: [String] = []
+        for image in images {
+            let observations = try await perform(on: image)
+            guard let doc = observations.first?.document else {
+                throw KankodaError.noDocumentFound
+            }
+            let texts = doc.paragraphs.map { $0.transcript }
+            results.append(contentsOf: texts)
+        }
+        return results
     }
 }
 #endif